sstable: verify components linkability
Verify that all components can be linked.

When fs.protected_hardlinks is enabled,
files must be either owned by the process uid
or have both read and write access so they
could be (hard) linked.

We don't go into these details but rather we
just try to link the components to a temporary name
and remove it, if succeeded.

Fixes #3117

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>