memtable: mc: fix updating encoding stats of collection tombstones
When the row flag has_complex_deletion is set, some collection columns may have
deletion tombstones and some may not. That said, the latter also need
to be written so we need to account for their values as well.

Currently the memtable code does not update encoding stats at all for live
collection cells in this case.

Fixes #4035

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>