memtable: mc: fix updating encoding stats of collection tombstones
When a row has complex deletion time, some collection columns may have
deletion tombstones and some may not. That said, the latter also need
to be written if the the row flags contain has_complex_deletion so we
need to account for their values as well.

Currently the memtable code does not update encoding stats at all for "live" tombstones.

Fixes #4035

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>